requirements.txt
crewai
crewai-tools
langchain-community
duckduckgo-search
wikipedia
requests
beautifulsoup4
python-dotenv
